wgengine/filter: add and clean up documentation.
